Moved some of the logic in the service class and added exception to check if User is deleted

+++ Forum Project/SportForumGroup10/src/main/java/com/telerikacademy/web/sportforumgroup10/services/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.telerikacademy.web.sportforumgroup10.services;
 
+import com.telerikacademy.web.sportforumgroup10.exceptions.AuthorizationException;
 import com.telerikacademy.web.sportforumgroup10.exceptions.EntityDuplicateException;
 import com.telerikacademy.web.sportforumgroup10.exceptions.EntityNotFoundException;
 import com.telerikacademy.web.sportforumgroup10.models.User;
@@ -12,6 +13,7 @@ import java.util.List;
 
 @Service
 public class UserServiceImpl implements UserService {
+    private static final String ERROR_MESSAGE = "You are not authorized";
     private final UserRepository userRepository;
 
     @Autowired
@@ -21,26 +23,39 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public List<User> getAllUsers() {
+
         return userRepository.getAllUsers();
     }
 
+    //TODO maybe no need only admins to search by id, firstName, email, username
+    // as it is mentioned in the requirements admin section
     @Override
-    public User getById(int id) {
+    public User getById(int id, User user) {
+        checkAccessPermissionId(id, user);
         return userRepository.getById(id);
     }
 
     @Override
-    public User getByFirstName(String firstName) {
+    public User getByFirstName(String firstName, User user) {
+
+        checkAccessPermissionString(firstName, user);
         return userRepository.getByFirstName(firstName);
     }
 
     @Override
-    public User getByEmail(String email) {
+    public User getByEmail(String email, User user) {
+        checkAccessPermissionString(email, user);
         return userRepository.getByEmail(email);
     }
 
     @Override
-    public User getByUsername(String username) {
+    public User getByUsername(String username, User user) {
+        checkAccessPermissionString(username, user);
+        return userRepository.getByUsername(username);
+    }
+
+    @Override
+    public User getByUsernameAuthentication(String username) {
         return userRepository.getByUsername(username);
     }
 
@@ -49,18 +64,19 @@ public class UserServiceImpl implements UserService {
         boolean duplicateUserExists = true;
         try {
             userRepository.getByUsername(user.getUsername());
-        } catch (EntityNotFoundException e){
+        } catch (EntityNotFoundException e) {
             duplicateUserExists = false;
         }
-        if (duplicateUserExists){
+        if (duplicateUserExists) {
             throw new EntityDuplicateException("User", "username", user.getUsername());
         }
-            return userRepository.create(user);
+        return userRepository.create(user);
     }
 
     @Override
     public User update(User user) {
-        User userToUpdate = getById(user.getId());
+        checkAccessPermissionId(user.getId(), user);
+        User userToUpdate = getById(user.getId(), user);
         userToUpdate.setFirstName(user.getFirstName());
         userToUpdate.setLastName(user.getLastName());
         userToUpdate.setEmail(user.getEmail());
@@ -69,7 +85,23 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User delete(int id) {
+    public User delete(int id, User userModifier) {
+        checkAccessPermissionId(id, userModifier);
         return userRepository.delete(id);
+    }
+
+    private void checkAccessPermissionId(int id, User requestingUser) {
+        if (!requestingUser.isAdmin()) {
+            if (requestingUser.getId() != id) {
+                throw new AuthorizationException(ERROR_MESSAGE);
+            }
+        }
+    }
+    private void checkAccessPermissionString(String input, User requestingUser) {
+        if (!requestingUser.isAdmin()) {
+//            if (!requestingUser.getUsername().equals(input)) {
+                throw new AuthorizationException(ERROR_MESSAGE);
+            }
+
     }
 }
